Initilize booking Status

+++ src/main/java/com/jh/hotelbookingmanagement/repository/BookingStatusRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookingStatusRepository extends JpaRepository<BookingStatus, Integer> {
+public interface BookingStatusRepository extends JpaRepository<BookingStatus, Long> {
 }
